requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
twilio>=8.10.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-multipart>=0.0.6
langchain>=0.0.350
openai>=1.3.0 